Create a new context provider that provides other contexts and remove useNavigation from FormContext
The reason to create a new context provider that returns all the contexts is because I have to use multiple contexts and this way makes it easier to maintain while rendering same children which is react router's elements in this case. Also the reason I removed useNavigation is because it can only be used for router elements rather than router itself. So it would give error in webpage.