Move input fields to ProductForm.jsx and delete useFormClose hook
The reason to move input fields in ProductForm is that the same fields could be used in product edit form. So, if i want to use either add or edit, i dont have to write all input fields again. Also I deleted useFormClose hook and decided to use the functions provided by useFromContext because the once the form is closed, the states have to be changed globally in form component as well as profile or other component where the form is used. the useFormClose hook won't update the states globally unlike contexts