Fix loading UI
While the product is loading in product page, the user can still navigate to other links.
But if an admin adds, updates or deletes product, until the request is not completed, they cannot nvaigate to other links.
The reason to do this was to prevent some bugs or glitches that might come up when request is still being handled by the server but the admin decides to go to another page which might cancel all the request and cause unexpected errors.