Fix navigation glitch after editing/deleting product
After editing the product, when admin clicks back button, instead of
navigating to previous page it would remain in the same product page even
after clicking back button multiple times.
The reason was fetching the item everytime single product page is loaded.
The data was fetched because of reFetchData being true after editing the product's information.
Also after deleting the product, it would navigate to product page which
is correct but if user tries to go to another product's page,
it would keep navigating back to product page and would not let user stay on single product's page.
The reason was because isProductDeleted would be true after product is
deleted but was never set back to false.