Use constant value and optimise fetchItem function
Used constant values instead of magic strings for fetchType parameter for fetchItems function.
This is because we should avoid magic strings as much as possible as it can be difficult to
maintain and debug code with random string values.
Another reason is if you have to change the value of one string, you should change value of every
strings. However, with one centralised value it will now be easier to debug and maintain code.